+++ lib/avromatic/model.rb
@@ -1,5 +1,5 @@
 require 'avromatic/model/builder'
-require 'avromatic/model/decoder'
+require 'avromatic/model/message_decoder'
 require 'avromatic/model/type_registry'
 
 module Avromatic